pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localrag"
version = "0.1.0"
description = "Locally indexes git repo and runs llm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colored>=2.3.1",
    "gitpython>=3.1.45",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-embeddings-ollama>=0.8.4",
    "llama-index-llms-ollama>=0.9.0",
    "llama-index-readers-file>=0.5.5",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "qdrant-client>=1.16.1",
    "tqdm>=4.67.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-language-pack>=0.13.0",
]

[dependency-groups]
dev = [
    "prek>=0.2.21",
    "pytest>=9.0.2",
]

[project.scripts]
localrag = "localrag.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/localrag"]
